utils: Move calculate_merkle_root to utils.
